+++ Changelog.md
@@ -9,8 +9,13 @@ Until then, its only a log of what I've changed to prepare update info on Nexus.
 ----
 
 # 2023.08.20 - 1.1.0
+##  CLA SSE - itself
 - Fixed: tbbmalloc / low ram - now share same text (partialy)
 - Fixed: regex handling for certain SKSE occourences/detections
+
+## Mod Issues
+- Added: (basic) Check for missing masters
+
 
 
 # 2023.07.10 - 1.0.1